Extract NoBareUrls
